% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdat.R
\name{simple.datf}
\alias{simple.datf}
\title{Add together two numbers}
\usage{
simple.datf(param)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
